validator: allow signs to have a single state
Until now the validator warned if anything was given in *:states for a signal
that has *:form=sign. However there instances of a signal with only one state
which is then given as only a sign (e. g. the German Bü0 signal), so only warn
if multiple states are given.