signals: show German Ks main signals if they have additional states
Allow either "off" or "DE-ESO:kennlicht" as additional states. It is as well
permitted to add "?" in case those additional states are possible, but unknown.

This is only done for new style Ks-signals (i.e. with DE-ESO: prefix) as the
old ones will likely have the old style Kennlicht tagging.